+++ new/parse_c_type.h
@@ -3,9 +3,9 @@
 
 typedef void *_cffi_opcode_t;
 
-#define _CFFI_OP(opcode, arg)   (_cffi_opcode_t)(opcode | (((unsigned long)(arg)) << 8))
-#define _CFFI_GETOP(cffi_opcode)    ((unsigned char)(unsigned long)cffi_opcode)
-#define _CFFI_GETARG(cffi_opcode)   (((unsigned long)cffi_opcode) >> 8)
+#define _CFFI_OP(opcode, arg)   (_cffi_opcode_t)(opcode | (((uintptr_t)(arg)) << 8))
+#define _CFFI_GETOP(cffi_opcode)    ((unsigned char)(uintptr_t)cffi_opcode)
+#define _CFFI_GETARG(cffi_opcode)   (((uintptr_t)cffi_opcode) >> 8)
 
 #define _CFFI_OP_PRIMITIVE       1
 #define _CFFI_OP_POINTER         3
